+++ Changelog.md
@@ -10,6 +10,7 @@
   * Added channels concept. Write only transports.
   * Twitter channel added.
   * Notification support for channels.
+  * Ybot memory json validating fix.
 
 ## Ybot-0.3.3 --> released